% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{calculate_items_per_manifest}
\alias{calculate_items_per_manifest}
\title{Call `manifest_download` Schematic endpoint for a given set of manifests and count the number of items in each manifest.}
\usage{
calculate_items_per_manifest(
  get_all_manifests_out,
  asset_view,
  input_token,
  base_url
)
}
\arguments{
\item{get_all_manifests_out}{output from `get_all_manifests()`}

\item{asset_view}{ID of view listing all project data assets. For example, for Synapse this would be the Synapse ID of the fileview listing all data assets for a given project.(i.e. master_fileview in config.yml)}

\item{input_token}{Synapse PAT}

\item{base_url}{Base URL of schematic API}
}
\description{
Call `manifest_download` Schematic endpoint for a given set of manifests and count the number of items in each manifest.
}
