% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{dt_replace_na}
\alias{dt_replace_na}
\title{NA replacement - datatable custom JS}
\usage{
dt_replace_na(col_index, na_replacement)
}
\arguments{
\item{df}{A dataframe with the columns `Contributor`, `Dataset_Name`, `Dataset_Type`, `Num_Items`, `Release_Scheduled`, `Embargo`, `Standard_Compliance`, `QC_Compliance`,`PHI_Detection_Compliance`, `Access_Controls_Compliance`, `Data_Portal`, `Released`, `past_due`}

\item{parsed_config}{updated config output by `parse_config()`}
}
\description{
NA replacement - datatable custom JS
}
