% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{generate_data_flow_manifest}
\alias{generate_data_flow_manifest}
\title{Generate a data flow status manifest. Fills in the component, contributor, data type, and dataset name columns.}
\usage{
generate_data_flow_manifest(
  storage_project_id,
  config,
  contributor = NULL,
  testing = FALSE
)
}
\description{
Generate a data flow status manifest. Fills in the component, contributor, data type, and dataset name columns.
}
