% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{get_na_replace_def}
\alias{get_na_replace_def}
\title{Parse config to get na replacement definitions with custom javascript. Outputs a list.}
\usage{
get_na_replace_def(config, prepped_dataframe)
}
\arguments{
\item{config}{datatable_dashboard_config.json as a datatable (`jsonlite::read_json("inst/datatable_dashboard_config.json")`)}

\item{prepped_dataframe}{dataframe output by `prep_df_for_dash()`}
}
\description{
Parse config to get na replacement definitions with custom javascript. Outputs a list.
}
