% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifeset_date_to_string}
\alias{manifeset_date_to_string}
\title{Convert date columns from date to string}
\usage{
manifeset_date_to_string(manifest)
}
\description{
Convert date columns from date to string
}
