% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_date_to_string}
\alias{manifest_date_to_string}
\title{Convert date columns from date to string}
\usage{
manifest_date_to_string(manifest)
}
\arguments{
\item{manifest}{a data flow status manifest.}
}
\description{
Convert date columns from date to string
}
