% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_generate}
\alias{manifest_generate}
\title{schematic rest api to generate manifest}
\usage{
manifest_generate(
  title,
  data_type,
  dataset_id,
  base_url = "https://schematic.dnt-dev.sagebase.org",
  schema_url =
    "https://raw.githubusercontent.com/ncihtan/data-models/main/HTAN.model.jsonld",
  oauth = "true",
  use_annotations = "false"
)
}
\arguments{
\item{title}{Name of dataset}

\item{data_type}{Type of dataset}

\item{dataset_id}{Synapse ID of existing manifest}

\item{base_url}{URL to schematic API endpoint}

\item{schema_url}{URL to a schema jsonld}

\item{oauth}{true or false STRING passed to python}

\item{use_annotations}{true or false STRING passed to python}
}
\value{
a URL to a google sheet
}
\description{
schematic rest api to generate manifest
}
