% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_string_to_date}
\alias{manifest_string_to_date}
\title{Convert date columns from string to date}
\usage{
manifest_string_to_date(manifest)
}
\arguments{
\item{manifest}{A data flow status manifest.}
}
\description{
Convert date columns from string to date
}
