% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{parse_config}
\alias{parse_config}
\title{Convert a vector of TRUE/FALSE to icon html}
\usage{
parse_config(config, df)
}
\arguments{
\item{config}{datatable_dashboard_config.json as a datatable (`jsonlite::read_json("inst/datatable_dashboard_config.json")`)}

\item{df}{dataframe to be turned into a dashboard}
}
\description{
Convert a vector of TRUE/FALSE to icon html
}
