% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{prep_manifest_dash}
\alias{prep_manifest_dash}
\title{Prepare a dataframe for a dashboard style datatable}
\usage{
prep_manifest_dash(df, past_due_col, config)
}
\arguments{
\item{df}{A dataframe with the columns `Contributor`, `Dataset_Name`, `Dataset_Type`, `Num_Items`, `Release_Scheduled`, `Embargo`, `Standard_Compliance`, `QC_Compliance`,`PHI_Detection_Compliance`, `Access_Controls_Compliance`, `Data_Portal`, `Released`, `past_due`}

\item{config}{Config for datatable dashboard module in `inst/datatable_dashboard_config.json`}
}
\description{
Prepare a dataframe for a dashboard style datatable
}
