% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{storage_dataset_files}
\alias{storage_dataset_files}
\title{/storage/dataset/files}
\usage{
storage_dataset_files(
  asset_view,
  dataset_id,
  input_token,
  file_names = list(),
  full_path = FALSE,
  base_url = "https://schematic.dnt-dev.sagebase.org"
)
}
\arguments{
\item{asset_view}{synapse ID of master file view.}

\item{dataset_id}{synapse ID of a storage dataset.}

\item{input_token}{synapse PAT}

\item{file_names}{a list of files with particular names (i.e. Sample_A.txt). If you leave it empty, it will return all dataset files under the dataset ID.}

\item{full_path}{Boolean. If True return the full path as part of this filename; otherwise return just base filename}

\item{base_url}{URL to schematic API endpoint}
}
\description{
/storage/dataset/files
}
