% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{storage_projects}
\alias{storage_projects}
\title{Get all storage projects the current user has access to}
\usage{
storage_projects(
  asset_view,
  input_token,
  url = "https://schematic.dnt-dev.sagebase.org/v1/storage/projects"
)
}
\arguments{
\item{asset_view}{synapse ID of master file view.}

\item{input_token}{synapse PAT}

\item{url}{URL to schematic API endpoint}
}
\description{
Get all storage projects the current user has access to
}
