% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{style_dashboard}
\alias{style_dashboard}
\title{Add custom styling to dashboard based on contents of config}
\usage{
style_dashboard(prepped_dataframe, parsed_config)
}
\arguments{
\item{parsed_config}{updated config output by `parse_config()`}

\item{df}{A dataframe prepared by `prep_df_for_dash()` with the columns `Contributor`, `Dataset_Name`, `Dataset_Type`, `Num_Items`, `Release_Scheduled`, `Embargo`, `Standard_Compliance`, `QC_Compliance`,`PHI_Detection_Compliance`, `Access_Controls_Compliance`, `Data_Portal`, `Released`, `past_due`}
}
\description{
Add custom styling to dashboard based on contents of config
}
