% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_functions.R
\name{prep_df_for_dash}
\alias{prep_df_for_dash}
\title{Prepare a dataframe for a dashboard style datatable}
\usage{
prep_df_for_dash(dataframe)
}
\arguments{
\item{A}{dataframe with the columns `Contributor`, `Dataset_Name`, `Dataset_Type`, `Num_Items`, `Release_Scheduled`, `Embargo`, `Standard_Compliance`, `QC_Compliance`,`PHI_Detection_Compliance`, `Access_Controls_Compliance`, `Data_Portal`, `Released`, `past_due`}
}
\description{
Prepare a dataframe for a dashboard style datatable
}
